now our casino support Mt and BetterMt play modes to randomize numbers

+++ src/main/java/com/bida/casino/royal/domain/dto/ResultDTO.java
@@ -7,5 +7,5 @@ public class ResultDTO {
 
     private AccountDTO account;
     private String message;
-    private int realNumber;
+    private long realNumber;
 }
